:recycle: Refactor page
page
